Refactor: - Extract formatting utils into its own file - Rename poorly named file to ParsedTemplate.kt - Review method visiblity - Render the amount by constructing a list of non null strings
